manage received file descriptors in a better way
Passing the ownership of file descriptors from a caller to a callee is
always a fragile thing to do as exceptions may get in the way. We opt
for a different model now. File descriptors are passed in a special
object of type OwnedFileDescriptors and the callee has to signal their
completion by returning a future or simply returning. The
OwnedFileDescriptors allows taking over ownership of individual file
descriptors voluntarily and the caller will ensure that all remaining
file descriptors will be closed.