ForeignVarlinkType: allow overriding the varlink type string
Doing so allows customizations beyond what is designed into this
library. If the library happens to fail supporting a use case, a user
may use a ForeignVarlinkType to pass through an object without
modifications and specify their own type string in the interface
description. The default of using "object" remains.