rewrite varlinkmethod decorator to better handle generators
In a varlink method call setting the more flag, each reply must indicate
whether there is a subsequent reply. This is not how Python generators
operate. When we receive one element from a generator, there may be
more, but it may also be the last one. We cannot tell, but varlink needs
to know.

There are now three ways to accomplish this and the user of the
varlinkmethod decorator gets to choose. One way is to wrap each result
in an AnnotatedResult object that indicates whether there will be
another one. This approach is very much like any other varlink
implementation does it, but it also is fairly annoying. Another approach
is delaying the generator by one element. That way, the previous element
can always indicate whether there will be another one as the next one
has already been consumed on the Python side. The last option involves
raising the last reply (if any) via a LastResult exception. No delay is
imposed, but the generator is assumed not never raise StopIteration or
AsyncStopIteration. It can be infinite or terminate via raising
LastResult.

Thanks to Enrico Zini for collaborating on this API approach.