refactor VarlinkMethodSignature.parameter_types
parameter_types and return_type were inconsistent on the Python side,
but quite similar on the varlink side. Since we now have a
ConversionError class, we can turn parameter_types into an
ObjectVarlinkType and still identify the location of a possible
conversion failure. We can also better express optional and mandatory
parameters now.

This is a breaking change, but there hopefully are no users yet.