refactor the lifetime management of file descriptor arrays
The life time of OwnedFileDescriptors was previously bound to a returned
future. The same future was used for backpressure. This way of managing
lifetime proved to be difficult to handle, so we now try a different
approach.

The new FileDescriptorArray (renamed) has a quite short lifetime, but a
consumer can extend its lifetime in a way similar to reference counting
- by attaching futures. When passed to a function, an array is collected
when the function returns unless it requests an extension. In cases
where the array is used in call_soon, such an extension becomes
required.

To avoid tedious lifetime management for the common case where there are
no file descriptors, a None value may be passed instead. Even though
this adds a conditional in a few places, it removes the need for
creating quite some futures.