conversion: fix optional parameters
We were applying two layers of optional. Once when a parameter is typed
"| None" and once when it has a default. The OptionalVarlinkType c'tor
does not like being nested and fails. Arguably, having a default does
not make a parameter optional in the OptionalVarlinkType sense as the
default may be something other than None. If it happens to be None, then
the type will already have expressed "| None". Therefore, removing the
second layer of OptionalVarlinkType is right.