VarlinkInterfaceProxy: turn error replies into exceptions
Before this change, an error reply was not recognized and instead it
would be decoded using the type information and most likely fail with a
ConversionError as a result. Instead, raise a GenericVarlinkErrorReply
now. This is not perfect as we might want individual error classes for
better matching, but getting some VarlinkErrorReply certainly is an
improvement as it lets a user see an actual error name.

Fixes: e5ca75915297 ("add VarlinkClientProtocol and VarlinkInterfaceProxy classes"