VarlinkProtocol: consume futures returned from request_received
When request_received returns a future, we may never inspect its result
as we're only interested in delaying processing until it completes.  As
a result, the asyncio stack may log a warning about its result never
being retrieved. It may also not log the warning when the garbage
collector does not get around to freeing the object. Instead, we now
explicitly retrieve the result and if it happens to be an exception, log
it immediately.