VarlinkServerProtocol: allow call_received fail the returned future
The meaning of the future returned from VarlinkServerProtocol wasn't
particularly well defined. It was forwarded to the VarlinkProtocol to
implement backpressure. The handling of a VarlinkErrorReply was
inconsistent in this regard. While it was turned into a reply when
raised directly from call_received, it wasn't when the VarlinkErrorReply
was set as exception on the returned future. This is changed now.
Additionally, other exceptions are now logged and the returned future
for backpressure never sets an exception.