VarlinkTransport: fail queued messages in correct order
When encountering an error while sending, we would previously fail all
queued messages before the message that failed to send. They should be
failed in submission order. Many other callers of _close_sender initiate
the call after verifying that the queue is empty and therefore do not
need to fail the queue.