Rework the FileDescriptor type
A fundamental problem arises when trying to return a socket.socket
object from a varlink method in a server context. The expected Python
type is a FileDescriptor, but it does not keep a reference to the socket
object and merely copies its fileno(). Once the socket goes out of
scope, it may be garbage collected and the FileDescriptor may become
invalid.

The returned object should include a reference to the socket object to
keep it alive. Hence the FileDescriptor type is changed to become a
regular object not inheriting from int. For now it continues to be used
in an immutable way.

This change propagates to all code that interfaces submits and receives
file descriptors. Objects now have to be wrapped in FileDescriptor
objects.