Change lifetime management of file descriptors
The FileDescriptorArray no longer looses its take method, so you may no
longer extract a FileDescriptor from it. Instead, the FileDescriptor's
take method can be used.

fromjson conversion no longer takes FileDescriptor. Therefore the file
descriptors contained in converted objects are not considered observers
of file descriptors rather than owners. The FileDescriptorArray remains
the owner.

On the server side, the FileDescriptorArray now remains valid until the
user-defined method returns. Prior, methods were responsible for closing
contained file descriptors.

On the client side, the VarlinkInterfaceProxy now checks whether the
return type contains a FileDescriptor somewhere and wraps its result in
a context manager if that happens to be the case.

Both the server side and the client side changes are breaking changes
for consumers.

Much thanks to Enrico Zini for coming up with the idea of using context
managers in this way. There remains a small risk of ignoring the context
manager and thereby never closing the managed file descriptors.