Do not return negative from FileDescriptor.fileno()
The behavior of fileno is not consistent in Python. A closed open(...)
object raises a ValueError whereas a socket.socket returns -1 after
closing. To ease use, ensure that FileDescriptor.fileno() always returns
a non-negative number or raises ValueError.

The choice is ValueError rather than negative numbers, because that
reduces the risk of accidental use of a bad file descriptor.