rework file descriptor handling again
The FileDescriptor class gains a release method. Releasing an owned
(should_close=True) FileDescriptor closes it whereas it does nothing for
an observed one (should_close=False).

This allows refactoring FileDescriptorArray into dealing with a mixture
of owned and unowned FileDescriptor objects. Consequently, drop the
"close" alias of "destroy" as it would be misleading.

The FileDescriptorArray also now enforces a uniqueness constraint. When
receiving file descriptors over a socket, this is automatically met.

With these changes and an "add" method, a FileDescriptorArray becomes
suitable for FileDescriptorVarlinkType.tojson where the out-of-band data
changes its type from a simple list[int] to a FileDescriptorArray.
Notably, this allows doing lifetime management of file descriptors
passed to tojson.

The client and server protocol implementations as well as the conversion
test mainly follow this change. Notably though, the FileDescriptorArray
populated by tojson is being lifetime extended with the future returned
from send_reply on the server side and this prevents premature closing
of returned file descriptors.

A new test case validates the desired behavior.

From an end user point of view, there are four main use cases.

1. Passing file descriptors to a client proxy. These should be passed as
   observed (retaining the should_close=False default) and remain valid
   until the method call returns.

2. Receiving file descriptors from a client proxy. These remain wrapped
   in a context manager controlling their lifetime.

3. Receiving file descriptors from a remote call. Their lifetime remains
   the responsibility of the library unless they are being "take()"n.

4. Returning file descriptors to a remote call. Here, the implementer
   can precisely control which of the file descriptors are to be closed
   by the library by setting or clearing the should_close argument.

Again thanks are due to Enrico Zini for discussing the user facing
semantics and thus shaping them.